requirements.txt
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
Deprecated==1.2.13
et-xmlfile==1.1.0
firebase==3.0.1
gcloud==0.18.3
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.4
jwcrypto==1.4.2
oauth2client==4.1.3
openpyxl==3.0.10
protobuf==4.21.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pyparsing==3.0.9
Pyrebase4==4.5.0
python-jwt==4.0.0
requests==2.28.1
requests-toolbelt==0.10.0
rsa==4.9
six==1.16.0
sseclient==0.0.27
urllib3==1.26.12
wrapt==1.14.1
python==3.10.6